Add docker-compose file and update app.js
Added a docker-compose file to simplify the environment setup process for development using a Node.js 14 container and a strategy for cloning, building and running the project. This will automate the process of setting up and initializing the database locally for faster development.

Also, updated app.js to create a new route ("/stinkypoopoohead.xml") that returns XML of the CiscoIPPhoneImageFile of a URL. This is part of enhancing the interactive client responsiveness.