Update docker-compose to simplify service setup
The changes involve merging some redundant steps and services in the docker-compose.yml file. This was done to simplify the setup of the app by combining the `npm install` and `npm start` commands together. Additionally, services such as 'wait-for-git', 'install-packages', and 'wait-for-node' were removed as they were no longer necessary due to the changes made to the command section of the services. The 'github-clone' service had its command updated to reflect the correct repository that is to be cloned.