"Update Node version and add Dockerfile"
Updated the Node version in docker-compose.yml from 14 to 18-alpine3.18 for better performance and added a new Dockerfile. The Dockerfile includes instructions to set up the application's environment, install app dependencies, copy the necessary files into the working directory, and run the application on port 3000. The node version in the Docker file has also been updated to match the version in docker-compose file for consistency.