Change volume binding path in docker-compose.yml
The volume binding path in the docker-compose.yml file was updated from './app:/app' to '/ipps:/ipps'. This change updates the application file path that Docker will mount to the directory path inside the Docker containers. The reason for the change is to align the project's Docker configuration to work with a different path, specifically the '/ipps' path. This ensures that the application will work correctly with the new directory structure.